moving step to its own step file
